# specs/purchase-kaikei/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: purchase-kaikei API（standalone v2）
  version: 2.0.0
  description: FEAT-PURCHASEKAIKEI — CT-API-PURCHASEKAIKEI-001..006。会計機能（仕訳・試算表・元帳）は内蔵エンジンで提供し、外部 API 連携はない。
paths:
  /api/purchases:
    get:
      summary: 申請一覧（状態フィルタ可）
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [submitted, approved, rejected, ordered, received]
      responses:
        "200": { description: OK }
    post:
      summary: 購買申請登録（部門付き）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [item, qty, unitPrice, requester]
              properties:
                item: { type: string }
                qty: { type: integer, minimum: 1 }
                unitPrice: { type: number, minimum: 0 }
                requester: { type: string }
                department: { type: string, enum: [D10, D20, D90], default: D90 }
                note: { type: string }
      responses:
        "201": { description: 作成（状態=submitted） }
        "400": { description: 必須項目不足・部門不正 }
  /api/purchases/{id}:
    get:
      summary: 詳細照会（履歴・仕訳参照含む）
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK }
        "404": { description: 未存在 }
  /api/purchases/{id}/approve:
    post:
      summary: 承認（1 段）
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: 状態→approved }
        "409": { description: invalid_transition }
  /api/purchases/{id}/reject:
    post:
      summary: 却下（comment 必須）
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [comment]
              properties:
                comment: { type: string }
      responses:
        "200": { description: 状態→rejected }
        "400": { description: comment 不足 }
        "409": { description: invalid_transition }
  /api/purchases/{id}/resubmit:
    post:
      summary: 再申請（rejected → submitted・履歴保持）
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: 状態→submitted }
        "409": { description: invalid_transition }
  /api/purchases/{id}/order:
    post:
      summary: 発注（approved のみ・発注番号採番）
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: 状態→ordered }
        "409": { description: invalid_transition }
  /api/purchases/{id}/receive:
    post:
      summary: 検収（ordered のみ・分割可）— 支払予定計上 + 仕訳 1 件自動計上（借 仕入高 5110 / 貸 買掛金 2110）
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                receivedQty: { type: integer, minimum: 1 }
                receivedAt: { type: string, format: date }
                comment: { type: string }
      responses:
        "200": { description: "検収成功。purchase と payable を返す。purchase.journal に entry id を記録" }
        "400": { description: 数量不正・残数量超 }
        "409": { description: invalid_transition }
  /api/payables:
    get:
      summary: 支払予定一覧（状態フィルタ可）
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [scheduled, paid] }
      responses:
        "200": { description: OK }
  /api/payables/{id}/pay:
    post:
      summary: 支払実行（scheduled のみ）— 仕訳 1 件自動計上（借 買掛金 2110 / 貸 普通預金 1120）
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: 支払成功。entry id を記録 }
        "404": { description: 未存在 }
        "409": { description: invalid_transition（二重支払含む） }
  /api/accounting/accounts:
    get:
      summary: 科目一覧（内蔵マスタ・11 科目）
      responses:
        "200": { description: OK }
  /api/accounting/trial-balance:
    get:
      summary: 残高試算表（内蔵エンジンが仕訳台帳から計算）
      parameters:
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
        - { name: department, in: query, schema: { type: string } }
      responses:
        "200": { description: "OK — 11 科目すべて（残高 0 を含む）・貸借一致" }
        "400": { description: 期間不正 }
  /api/accounting/ledger/{account_code}:
    get:
      summary: 総勘定元帳（内蔵エンジンが仕訳台帳から計算）
      parameters:
        - { name: account_code, in: path, required: true, schema: { type: string } }
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses:
        "200": { description: "OK — 日付昇順・残高列（繰越）付き" }
        "404": { description: 科目未存在 }